Add build for windows
Implemented missing windows functionality (Timer and GenerateRandomSeed).
Changed paho to compile with vc++ 2017 (removed zero-sized array in MQTT::Client class, fixed FP::FPtrDummy to be an empty instead of undefined class).
Supressed warnings in milagro-crypto-c and paho MQTTPacket.
Fixed warnings in sdk code.

+++ tests/iot_client/flags.cpp
@@ -1,6 +1,7 @@
 #include "flags.h"
 #include <iostream>
 #include <string.h>
+#include <algorithm>
 
 using std::cout;
 using std::endl;
@@ -99,7 +100,7 @@ void Flags::ApplyDefaults()
 
 static std::string Align(const std::string& s, size_t len)
 {
-    int diff = len - s.length();
+    int diff = static_cast<int>(len - s.length());
     if (diff <= 0)
     {
         return s;
